refactor(http2): split src/server.ts into per-concern modules (1453 → 49 + 4)
Applies the proven webgl/process/tls per-concern split pattern to the
http2 server-side surface. `src/server.ts` shrinks from 1453 LoC to a
49-LoC thin composer that re-exports from `src/server/`; implementation
is redistributed across four files:

  - server/request.ts        (76 LoC)  — Http2ServerRequest Readable
  - server/session.ts        (140 LoC) — ServerHttp2Session +
                                         _allocatePushId +
                                         _buildPushPromise (native bridge)
  - server/response.ts       (811 LoC) — Http2NativeBackend iface +
                                         Http2ServerResponse +
                                         ServerHttp2Stream + the
                                         _respondFromFD /
                                         _makeDetachedSoupMessage helpers
                                         (Response↔Stream kept together
                                         due to tight coupling)
  - server/http2-server.ts   (452 LoC) — Http2Server + Http2SecureServer
                                         + _toPemString /
                                         _createTlsCertificate +
                                         _activeServers GC-guard Set

Public surface preserved verbatim: every export from `@gjsify/http2` /
`@gjsify/http2/server` works unchanged. 102 Node + 231 GJS tests still
green; type-check + library build clean. No new behaviour, no new tests
— straight refactor following the prior `@gjsify/process`,
`@gjsify/tls`, and `@gjsify/webgl` splits.

STATUS.md: new entry under Completed.
