+++ STATUS.md
@@ -398,6 +398,8 @@ Not yet implemented (but potentially relevant for GJS projects):
 
 ### Completed
 
+- ~~**`@gjsify/http2` — split `src/server.ts` into per-concern modules (1453 → 49 + 4 modules)**~~✓ — Applies the proven webgl/process/tls per-concern split pattern to the http2 server-side surface. `src/server.ts` shrinks from 1453 LoC to a 49-LoC thin composer that just re-exports from `src/server/`; the implementation is redistributed across four files: `server/request.ts` (76 LoC — `Http2ServerRequest` Readable side), `server/session.ts` (140 LoC — `ServerHttp2Session` with `_allocatePushId` + `_buildPushPromise` via the @gjsify/http2-native bridge), `server/response.ts` (811 LoC — `Http2NativeBackend` interface + `Http2ServerResponse` + `ServerHttp2Stream` + the `_respondFromFD`/`_makeDetachedSoupMessage` helpers — kept together because Response and Stream are tightly coupled), `server/http2-server.ts` (452 LoC — `Http2Server` + `Http2SecureServer` classes + `_toPemString`/`_createTlsCertificate` TLS helpers + the `_activeServers` GC-guard Set). All public surface preserved: same exports, same imports work unchanged from `@gjsify/http2`. 102 Node + 231 GJS http2 tests still green; type-check + library build clean. No new behaviour, no new tests — straight refactor.
+
 - ~~**`@gjsify/dom-elements` `ResizeObserver` polyfill: real bridge wiring (replaces stub)**~~✓ — `ResizeObserver` was a no-op stub: every method returned undefined, `observe()` never fired its callback. Excalibur.js 0.32's `Screen._setResolutionAndViewportByDisplayMode()` installs `new ResizeObserver(cb).observe(canvas.parentElement || document.body)` under `DisplayMode.FillContainer`; the stub broke that path, leaving rendered scenes stuck at their initial GTK allocation. **Fix:** real implementation in `packages/dom/dom-elements/src/resize-observer.ts`. `Element` gains `_onResize(cb): () => void` (subscribe + unsubscribe) and `_fireResizeSubscribers(width, height)` (internal dispatch); `notifyElementResize(target, width, height)` (new `notify-resize.ts`, also exported from the public surface) fires the target's own subscribers + walks `parentNode` chain firing ancestor subscribers — the latter is what makes Excalibur's `observe(canvas.parentElement)` pattern work, since the bridge resize signal hits the canvas, not its parent. Resize storms collapse into a single microtask-batched callback per observer (matches the spec's "deliver resize loop notifications" cycle). Each bridge that owns a polyfill DOM element now forwards its GTK `resize` signal: `WebGLBridge` (`Gtk.GLArea` → `_canvas`), `Canvas2DBridge` (`Gtk.DrawingArea` → `_canvas`), `VideoBridge` (`Gtk.Picture` → `_video`), `IFrameBridge` (`WebKit.WebView` → `_iframe`, **new resize listener**). 8 new specs in `packages/dom/dom-elements/src/resize-observer.spec.ts` (direct fire, ancestor walk, unobserve / disconnect, idempotent observe, multi-observer isolation, entry shape, microtask batching) — all 442 dom-elements tests pass on Node + GJS. Closes the PixelRPG map-editor `FillContainer` parity gap documented in the original report.
 
 - ~~**`css-as-string` plugin: restore npm-package `@import` resolution (esbuild → rolldown regression)**~~✓ — `@gjsify/rolldown-plugin-gjsify`'s `cssAsStringPlugin` calls lightningcss `bundleAsync` to resolve `@import` chains during `.css` → JS-string conversion. The old esbuild-based predecessor (`@gjsify/esbuild-plugin-css`, retired in the rolldown migration) walked `node_modules` for bare specifiers + honoured `package.json#exports`, so consumers could write `@import "@scoped/pkg/styles.css"` from any workspace CSS file. The npm-track of the new plugin invoked `bundleAsync` with no `resolver`, leaving lightningcss's default file-only walker — which only handles relative + absolute paths and otherwise hits ENOENT. Surfaced by the `pixel-rpg/map-editor` migration: `apps/maker-gjs/src/application.css` couldn't import `@pixelrpg/gjs/index.css` and required a 3-segment relative-path workaround. Fix: add a `cssBundleResolver` (~20 LoC) that distinguishes the three path shapes — absolute pass-through, relative resolved against the importer's dirname, bare specifier resolved via `createRequire(pathToFileURL(from).href).resolve(specifier)` so `node_modules` lookups + `exports` maps both work. The same resolver is now wired into `loadNpmBundler()`'s `bundleAsync` call. Native track (`@gjsify/lightningcss-native`) is unchanged — its Vala bridge signature takes a single `filename` + browserslist string and has no resolver hook; documented limitation, only relevant under GJS-runs-build (Phase D-3) which downstream monorepos don't hit yet. `tests/e2e/css-bundling/run.mjs` extended with a second test case that plants a fake `@css-fixture/shared` package under `node_modules` (with `package.json#exports` exposing `./styles.css`) and asserts (a) the bare-specifier @import resolves end-to-end, (b) the resolved CSS content is inlined into the bundled JS string, (c) relative + bare @imports cohabit in the same file. Unblocks the map-editor CSS workaround which can now be reverted in its next gjsify-cli bump.
